Membuat dialog konfirmasi hapus

+++ app/src/main/java/org/d3if3141/asessment2mobpro/ui/screen/DetailScreen.kt
@@ -64,6 +64,7 @@ fun DetailScreen(navController: NavHostController, id: Long? = null) {
     var jumlah by remember { mutableStateOf("") }
     var tenggat by remember { mutableStateOf("") }
 
+    var showDialog by remember { mutableStateOf(false)}
 
     val radioOptions = listOf(
         stringResource(id = R.string.tigabln),
@@ -120,7 +121,12 @@ fun DetailScreen(navController: NavHostController, id: Long? = null) {
                         )
                     }
                     if (id != null) {
-                        DeleteAction {
+                        DeleteAction { showDialog = true }
+                        DisplayAlertDialog(
+                            openDialog = showDialog,
+                            onDismissRequest = { showDialog = false }
+                        ) {
+                            showDialog = false
                             viewModel.delete(id)
                             navController.popBackStack()
                         }
